Run the HMPPS rename script (from the template)
