+++ applicationinsights.json
@@ -1,6 +1,6 @@
 {
   "role": {
-    "name": "hmpps-template-kotlin"
+    "name": "legacy-ppud-api"
   },
   "customDimensions": {
     "service.version": "${BUILD_NUMBER}"
