+++ docker-compose.yml
@@ -1,10 +1,10 @@
 version: "3"
 services:
-  hmpps-template-kotlin:
+  legacy-ppud-api:
     build:
       context: .
     network_mode: "host"
-    container_name: hmpps-template-kotlin
+    container_name: legacy-ppud-api
     ports:
       - "8081:8080"
     healthcheck:
