+++ helm_deploy/legacy-ppud-api/values.yaml
@@ -1,18 +1,18 @@
 ---
 generic-service:
-  nameOverride: hmpps-template-kotlin
+  nameOverride: legacy-ppud-api
 
   replicaCount: 4
 
   image:
-    repository: quay.io/hmpps/hmpps-template-kotlin
+    repository: quay.io/hmpps/legacy-ppud-api
     tag: app_version    # override at deployment time
     port: 8080
 
   ingress:
     enabled: true
     host: app-hostname.local    # override per environment
-    tlsSecretName: hmpps-template-kotlin-cert
+    tlsSecretName: legacy-ppud-api-cert
     path: /
 
   # Environment variables to load into the deployment
@@ -29,7 +29,7 @@ generic-service:
   #     [name of environment variable as seen by app]: [key of kubernetes secret to load]
 
   namespace_secrets:
-    hmpps-template-kotlin:
+    legacy-ppud-api:
       APPINSIGHTS_INSTRUMENTATIONKEY: "APPINSIGHTS_INSTRUMENTATIONKEY"
 
   allowlist:
@@ -41,4 +41,4 @@ generic-service:
     cloudplatform-live1-3: "35.177.252.54/32"
 
 generic-prometheus-alerts:
-  targetApplication: hmpps-template-kotlin
+  targetApplication: legacy-ppud-api
