+++ src/main/kotlin/uk/gov/justice/digital/hmpps/legacyppudapi/config/LegacyPpudApiExceptionHandler.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppstemplatepackagename.config
+package uk.gov.justice.digital.hmpps.legacyppudapi.config
 
 import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice
 import javax.validation.ValidationException
 
 @RestControllerAdvice
-class HmppsTemplateKotlinExceptionHandler {
+class LegacyPpudApiExceptionHandler {
   @ExceptionHandler(ValidationException::class)
   fun handleValidationException(e: Exception): ResponseEntity<ErrorResponse> {
     log.info("Validation exception: {}", e.message)
