+++ src/main/kotlin/uk/gov/justice/digital/hmpps/legacyppudapi/health/VersionOutputter.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.hmppstemplatepackagename.health
+package uk.gov.justice.digital.hmpps.legacyppudapi.health
 
 import org.slf4j.LoggerFactory
 import org.springframework.boot.context.event.ApplicationReadyEvent
