+++ src/main/resources/logback-spring.xml
@@ -25,7 +25,7 @@
         <appender name="logAppender" class="com.microsoft.applicationinsights.logback.ApplicationInsightsAppender"/>
     </springProfile>
 
-    <logger name="uk.gov.justice.digital.hmpps.hmppstemplatepackagename.HmppsTemplateKotlinKt" additivity="false"
+    <logger name="uk.gov.justice.digital.hmpps.legacyppudapi.LegacyPpudApiKt" additivity="false"
             level="DEBUG">
         <appender-ref ref="logAppender"/>
         <appender-ref ref="consoleAppender"/>
